+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fix Firefox manifest
+
 ## 1.5.0 - 2024 Jun 23
 - Use full quality SoundCloud covers
 - Update to Manifest V3